Factor out instance loading logic into a mixin; add load_instance option
addresses #193
